Rename pympipool to executorlib
