+++ .github/workflows/benchmark.yml
@@ -43,7 +43,7 @@ jobs:
         python tests/benchmark/llh.py process >> timing.log
         python tests/benchmark/llh.py thread >> timing.log
         mpiexec -n 4 python -m mpi4py.futures tests/benchmark/llh.py mpi4py >> timing.log
-        python tests/benchmark/llh.py pympipool >> timing.log
+        python tests/benchmark/llh.py executorlib >> timing.log
         python tests/benchmark/llh.py block_allocation >> timing.log
         cat timing.log
         python -m unittest tests/benchmark/test_results.py
