+++ .pre-commit-config.yaml
@@ -5,6 +5,6 @@ repos:
       - id: ruff
         name: ruff lint
         args: ["--select", "I", "--fix"]
-        files: ^pympipool/
+        files: ^executorlib/
       - id: ruff-format
         name: ruff format
