+++ executorlib/_version.py
@@ -50,9 +50,9 @@ def get_config() -> VersioneerConfig:
     cfg = VersioneerConfig()
     cfg.VCS = "git"
     cfg.style = "pep440-pre"
-    cfg.tag_prefix = "pympipool-"
-    cfg.parentdir_prefix = "pympipool"
-    cfg.versionfile_source = "pympipool/_version.py"
+    cfg.tag_prefix = "executorlib-"
+    cfg.parentdir_prefix = "executorlib"
+    cfg.versionfile_source = "executorlib/_version.py"
     cfg.verbose = False
     return cfg
