+++ executorlib/backend/cache_parallel.py
@@ -3,7 +3,7 @@ import sys
 
 import cloudpickle
 
-from pympipool.cache.shared import backend_load_file, backend_write_file
+from executorlib.cache.shared import backend_load_file, backend_write_file
 
 
 def main():
