+++ executorlib/backend/interactive_parallel.py
@@ -4,8 +4,8 @@ from os.path import abspath
 
 import cloudpickle
 
-from pympipool.interactive.backend import call_funct, parse_arguments
-from pympipool.shared.communication import (
+from executorlib.interactive.backend import call_funct, parse_arguments
+from executorlib.shared.communication import (
     interface_connect,
     interface_receive,
     interface_send,
