+++ executorlib/shared/communication.py
@@ -10,7 +10,7 @@ class SocketInterface(object):
     The SocketInterface is an abstraction layer on top of the zero message queue.
 
     Args:
-        interface (pympipool.shared.interface.BaseInterface): Interface for starting the parallel process
+        interface (executorlib.shared.interface.BaseInterface): Interface for starting the parallel process
     """
 
     def __init__(self, interface=None):
@@ -117,8 +117,8 @@ def interface_bootup(
 
     Args:
         command_lst (list): List of commands as strings
-        connections (pympipool.shared.interface.BaseInterface): Interface to start parallel process, like MPI, SLURM or
-                                                                Flux
+        connections (executorlib.shared.interface.BaseInterface): Interface to start parallel process, like MPI, SLURM
+                                                                  or Flux
         hostname_localhost (boolean): use localhost instead of the hostname to establish the zmq connection. In the
                                       context of an HPC cluster this essential to be able to communicate to an
                                       Executor running on a different compute node within the same allocation. And
@@ -130,7 +130,7 @@ def interface_bootup(
         prefix_path (str): path of the conda environment to initialize
 
     Returns:
-         pympipool.shared.communication.SocketInterface: socket interface for zmq communication
+         executorlib.shared.communication.SocketInterface: socket interface for zmq communication
     """
     if not hostname_localhost:
         command_lst += [
