+++ executorlib/shell/__init__.py
@@ -1,0 +1,7 @@
+from executorlib.shell.executor import SubprocessExecutor
+from executorlib.shell.interactive import ShellExecutor
+
+__all__ = [
+    SubprocessExecutor,
+    ShellExecutor,
+]
