--- pympipool/shell/__init__.py
@@ -1,7 +1,0 @@
-from pympipool.shell.executor import SubprocessExecutor
-from pympipool.shell.interactive import ShellExecutor
-
-__all__ = [
-    SubprocessExecutor,
-    ShellExecutor,
-]
