+++ tests/test_cache_executor_mpi.py
@@ -5,8 +5,8 @@ import unittest
 
 
 try:
-    from pympipool import FileExecutor
-    from pympipool.cache.shared import execute_tasks_h5, execute_in_subprocess
+    from executorlib import FileExecutor
+    from executorlib.cache.shared import execute_tasks_h5, execute_in_subprocess
 
     skip_h5io_test = False
 except ImportError:
