+++ tests/test_cache_executor_serial.py
@@ -4,11 +4,11 @@ from queue import Queue
 import shutil
 import unittest
 
-from pympipool.shared.thread import RaisingThread
+from executorlib.shared.thread import RaisingThread
 
 try:
-    from pympipool import FileExecutor
-    from pympipool.cache.shared import execute_tasks_h5, execute_in_subprocess
+    from executorlib import FileExecutor
+    from executorlib.cache.shared import execute_tasks_h5, execute_in_subprocess
 
     skip_h5io_test = False
 except ImportError:
