+++ tests/test_executor_backend_mpi.py
@@ -1,8 +1,8 @@
 import importlib.util
 import unittest
 
-from pympipool import Executor
-from pympipool.shared.executor import cloudpickle_register
+from executorlib import Executor
+from executorlib.shared.executor import cloudpickle_register
 
 
 skip_mpi4py_test = importlib.util.find_spec("mpi4py") is None
