+++ tests/test_executor_backend_mpi_noblock.py
@@ -1,7 +1,7 @@
 import unittest
 
-from pympipool import Executor
-from pympipool.shared.executor import cloudpickle_register
+from executorlib import Executor
+from executorlib.shared.executor import cloudpickle_register
 
 
 def calc(i):
