+++ tests/test_shared_input_check.py
@@ -1,6 +1,6 @@
 import unittest
 
-from pympipool.shared.inputcheck import (
+from executorlib.shared.inputcheck import (
     check_command_line_argument_lst,
     check_gpus_per_worker,
     check_threads_per_core,
